VideoDetail and FavoriteVideoDetail are pages now
-VideoDetail and FavoriteVideoDetail where changed from '/components' to '/pages' since they behavior as pages.

-Because those two actually use two shared almost the same code using components almost the same I decided to make some changes in the VideoInfo and RelatedVideos component to reuse the code I had already done in order to make this mentioned components easier to mantain since the logic of data they use now is handled by the page.

-The VideoInfo and VideoDetail components are now directly under '/components' in an effor to make clear that the components which are directly under the mentioned directory are always usable in a standalone way, so if needed they can be reused more often.

-Sidechange. Now the Layout component implements global style in the body, this will stop a white bakground color dividing the screen in some cases were the dark theme is used.
