+++ src/components/VideoFrame/VideoFrame.styles.js
