+++ src/pages/Home/Home.page.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
-import { VideosGrid } from '../../components/VideosGrid/VideosGrid.component';
-import { useFetchVideos } from '../../utils/hooks/useFetchVideos';
-import './Home.styles.css';
-import styled from 'styled-components';
+
+//hooks
 import { useGlobal } from '../../providers/Global.provider';
 
-const IconsInfoContainer = styled.div`
-  color: ${props => props.theme.iconsInfoContainer.color};
-`;
+//components
+import { VideosGrid } from '../../components/VideosGrid/VideosGrid.component';
+import { useFetchVideos } from '../../utils/hooks/useFetchVideos';
+
+//Styles
+import { Container, IconsInfoContainer } from './Home.styles';
 
 function HomePage() {
 
@@ -15,13 +16,12 @@ function HomePage() {
   const { videos, loading } = useFetchVideos(state.search);
 
   return (
-    <section className="homepage" >
+    <Container>
       {loading && <p>Loading</p>}
-      {!loading && <VideosGrid data={videos} location='/video'/>}  
-      
-    <IconsInfoContainer theme={state.theme}>Icons made by <a href="https://www.freepik.com" title="Freepik">Freepik</a> from <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a></IconsInfoContainer>
-    <IconsInfoContainer theme={state.theme}>Icons made by <a href="https://icon54.com/" title="Pixel perfect">Pixel perfect</a> from <a href="https://www.flaticon.es/" title="Flaticon">www.flaticon.com</a></IconsInfoContainer>
-    </section>
+      {!loading && <VideosGrid data={videos} location='/video'/>}
+      <IconsInfoContainer theme={state.theme}>Icons made by <a href="https://www.freepik.com" title="Freepik">Freepik</a> from <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a></IconsInfoContainer>
+      <IconsInfoContainer theme={state.theme}>Icons made by <a href="https://icon54.com/" title="Pixel perfect">Pixel perfect</a> from <a href="https://www.flaticon.es/" title="Flaticon">www.flaticon.com</a></IconsInfoContainer>
+    </Container>
   );
 }
