--- src/pages/NotFound/NotFound.styles.css
@@ -1,13 +1,0 @@
-.home-link {
-  font-size: 1rem;
-  text-decoration: none;
-  display: block;
-  text-align: right;
-  padding: 0.2rem 0.3rem;
-}
-
-.home-link::before {
-  content: '←';
-  padding-right: 0.2rem;
-  display: inline-block;
-}
