+++ src/pages/NotFound/NotFound.styles.js
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const Container = styled.section`
+    display: grid;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    text-align: center;
+    width: 100%;
+`;
+
+export const TitleLink = styled.h1`
+    text-decoration: underline;
+`;
+
+export const ImgContainer = styled.div`
+    width: 100%;
+    img{
+        border-radius: 50%;
+    }
+`;
