+++ src/pages/VideoDetail/VideoDetail.styles.js
