+++ src/utils/extractVideos.js
@@ -1,0 +1,9 @@
+export const extractVideos = (data) => {
+    return data.items.map(item=>{
+        console.log(data);
+        if(item.id.hasOwnProperty('videoId')){
+            item.id = item.id.videoId;
+        }
+        return item;
+    });
+}
