+++ src/utils/hooks/useGapi.js
@@ -5,7 +5,7 @@ export const useGapi = () => {
 
     const initGapi = async() =>{
         try{
-            await window.gapi.client.setApiKey(process.env.YOUTUBE_API_KEY);
+            await window.gapi.client.setApiKey(process.env.REACT_APP_YOUTUBE_API_KEY);
             await window.gapi.client.load('https://www.googleapis.com/discovery/v1/apis/youtube/v3/rest');
             setGapi(window.gapi);
         }catch(e){
